+++ Sources/BraintreePayPal/BTPayPalCheckoutRequest.swift
@@ -82,7 +82,7 @@ import BraintreeCore
     ///   - requestBillingAgreement: Optional: If set to `true`, this enables the Checkout with Vault flow, where the customer will be prompted to consent to a billing agreement
     ///   during checkout. Defaults to `false`.
     ///   - contactPreference: Optional: Preference for the contact information section within the payment flow. Defaults to `.none` if not set.
-    ///   - shouldOfferCredit: Optional: Offers PayPal Credit if the customer qualifies. Defaults to `false`. Only available with PayPal Checkout.
+    ///   - offerCredit: Optional: Offers PayPal Credit if the customer qualifies. Defaults to `false`.
     /// - Warning: This initializer should be used for merchants using the PayPal App Switch flow. This feature is currently in beta and may change or be removed in future releases.
     /// - Note: The PayPal App Switch flow currently only supports the production environment.
     public convenience init(
@@ -95,7 +95,7 @@ import BraintreeCore
         currencyCode: String? = nil,
         requestBillingAgreement: Bool = false,
         contactPreference: BTContactPreference = .none,
-        shouldOfferCredit: Bool = false
+        offerCredit: Bool = false
     ) {
         self.init(
             amount: amount,
@@ -105,7 +105,7 @@ import BraintreeCore
             currencyCode: currencyCode,
             requestBillingAgreement: requestBillingAgreement,
             userAuthenticationEmail: userAuthenticationEmail,
-            shouldOfferCredit: shouldOfferCredit
+            offerCredit: offerCredit
         )
         super.enablePayPalAppSwitch = enablePayPalAppSwitch
     }
@@ -127,7 +127,7 @@ import BraintreeCore
     ///   - recurringBillingDetails: Optional: Recurring billing product details.
     ///   - recurringBillingPlanType: Optional: Recurring billing plan type, or charge pattern.
     ///   - amountBreakdown: Optional: Breakdown of items associated to the total cost.
-    ///   - shouldOfferCredit: Optional: Offers PayPal Credit if the customer qualifies. Defaults to `false`. Only available with PayPal Checkout.
+    ///   - offerCredit: Optional: Offers PayPal Credit if the customer qualifies. Defaults to `false`.
     public init(
         amount: String,
         intent: BTPayPalRequestIntent = .authorize,
@@ -140,7 +140,7 @@ import BraintreeCore
         recurringBillingDetails: BTPayPalRecurringBillingDetails? = nil,
         recurringBillingPlanType: BTPayPalRecurringBillingPlanType? = nil,
         amountBreakdown: BTAmountBreakdown? = nil,
-        shouldOfferCredit: Bool = false
+        offerCredit: Bool = false
     ) {
         self.amount = amount
         self.intent = intent
@@ -157,7 +157,7 @@ import BraintreeCore
             recurringBillingDetails: recurringBillingDetails,
             recurringBillingPlanType: recurringBillingPlanType,
             userAction: userAction,
-            shouldOfferCredit: shouldOfferCredit
+            offerCredit: offerCredit
         )
     }
 
@@ -175,7 +175,7 @@ import BraintreeCore
             "intent": intent.stringValue,
             "amount": amount,
             "offer_pay_later": offerPayLater,
-            "offer_paypal_credit": shouldOfferCredit
+            "offer_paypal_credit": offerCredit
         ]
 
         let currencyCode = currencyCode != nil ? currencyCode : configuration.json?["paypal"]["currencyIsoCode"].asString()